Store a z-index id per-layer

+++ crates/editor/src/element.rs
@@ -3169,7 +3169,7 @@ pub struct CursorName {
     string: SharedString,
     color: Hsla,
     is_top_row: bool,
-    z_index: u8,
+    z_index: u16,
 }
 
 impl Cursor {
